+++ src/datatypes/flag.ts
@@ -1,0 +1,6 @@
+// 旗帜种类 flag
+export enum Flag {
+    None = "none", // 无
+    Tradeable = "tradeable", // 可交易
+    Forge = "forge" // 锻造
+}