+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
-- Alpha Ver. v0.0.3  -  Add `i18n`
-- Alpha Ver. v0.0.2  -  Add `spell cards` generation.
-- Alpha Ver. v0.0.1  -  `Minion Cards` generation.
+## Alpha Ver. v0.0.3
+- Add `i18n`
+- Add `Battlegrounds Minion` generation.
+- Add `Flag` of Minions such as `Tradeable` or `Forge`
+
+## Alpha Ver. v0.0.2
+Add `spell cards` generation.
+
+## Alpha Ver. v0.0.1
+`Minion Cards` generation.